Update Project.ipynb
The project includes a python package that make a request to an external API, retrieves data for a given cryptocurrency asset, and filters the results based on specified optional parameters including size, name of exchange, base asset and quote asset. Additionally, the function includes error handling and type hints to improve its readability and robustness.